bugfix: Address issues with generating/simulating daily policy system data
The first cut had a number of issues. There was an awkward method to
figure out how to dump the next day of data only and it didn't quite work.

The simple way to do it is just use the simulated system state next modified
date to differentiate what is new data to be dumped. This worked but the way
the copy job worked was different between type tables and all the other
tables as the type tables did not originally contain the header along with
the data. These were brought in to line with the other tables so only one
code path needed to be managed.

Finally, when tables had nothing new e.g the type tables, they were not
getting cleaned up, so added a check for files with only 1 row. When only
1 row, these were cleaned up as they just contain headers.