feat: Finished cdc to bronze functionality Data is now captured as a daily delta in sync with the upstream practice of dumping each day of data to flat file then uploading to the replica database.
This identifies the delta and pulls it out of the database. State
is now retained so the script knows which day it is up to and
will detect if a local "bronze" environment (e.g landing/raw)
does not exist.

The script is now also integrated properly with the rest of the
project in terms of secrets management and the database to use