bugfix: Policyforge replica data structures
The replica data structures did not support the delta process that
the policy generator follows. This change relaxes foreign key
constraints (which are no longer valid due to modified date not
being part of the relationship, but now defining what makes a
record unique across all tables in combination with the entity ID)
and updates the primary key on each table to reflect how the
system simulation works - which is to go back and modify a record
rather than generate a new one for each event that takes place.