feat: Add proof of concept dbt project for silver layer
This adds a DBT project which builds off a table that has been
manually loaded into the bronze layer and is therefore not currently
tracked in source control.

Some very basic tests are included and the project is integrated to
Dagster, which can kick off a dev or a prod run.

A little tinkering was needed to the Dagster workspace file to get
everything to play together. It's not ideal in terms of structure and
needs to be addressed at a later date but is functional