feat: Update CDC process
Updates CDC process to utilise new infra config file. The CDC file
now publishes output to blob storage in a format that interfaces with
the Snowflake structure nicely (each table has its own folder and each
delta is a timestamped file in that folder).

The process now pushes delta files to blob storage on run and rerefreshes
the external tables in Snowflake